+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 Changelog
 =========
 
+### 0.0.3 (2014-09-07)
+- add script to manage virtual box machines (vbox-mgr)
+- add supporting libs for above
+- add yaml as a requirement (after *salt* itself needs it :> )
+
 ### 0.0.2 (2014-09-02)
 - added changelog (yay!)
 - added data dir